feat: :building_construction: Relocate Table component
