+++ libs/dana/src/components/Table/DataTable.tsx
@@ -1,6 +1,6 @@
 import { ReactElement } from 'react';
 
-import { Props } from '../../../helpers';
+import { Props } from '../../helpers';
 
 export interface DataTableProps extends Props {
     children: ReactElement;
