+++ libs/dana/src/components/Table/Table.stories.tsx
@@ -1,7 +1,7 @@
 import { action } from '@storybook/addon-actions';
 
-import type { Story } from '../../../lib/storybook-emotion-10-fixes';
-import { asChromaticStory } from '../../../lib/story-intents';
+import type { Story } from '../../lib/storybook-emotion-10-fixes';
+import { asChromaticStory } from '../../lib/story-intents';
 import {
     DataTable,
     DataTableProps,
