+++ libs/dana/src/components/Table/TableToolbarSearch.tsx
@@ -1,9 +1,9 @@
 import { SearchIcon } from '@dana-icons';
 import React, { InputHTMLAttributes, useState } from 'react';
-import { Props } from '../../../helpers';
-import { Button } from '../../forms/button';
-import { Select, Option } from '../../forms/select';
-import { TextInput } from '../../forms/text-input';
+import { Props } from '../../helpers';
+import { Button } from '../forms/button';
+import { Select, Option } from '../forms/select';
+import { TextInput } from '../forms/text-input';
 import {
     tableToolbarSearch,
     tableToolbarSearchButton,
