--- libs/dana/src/components/data-display/index.ts
@@ -1,1 +1,0 @@
-export * from './table';
