+++ libs/dana/src/components/index.ts
@@ -5,7 +5,6 @@ export * from './text';
 export * from './title';
 
 export * from './forms';
-export * from './data-display';
 export * from './navigation';
 export * from './layout';
 
@@ -25,3 +24,4 @@ export * from './RadioGroup';
 export * from './Badge';
 export * from './Card';
 export * from './Divider';
+export * from './Table';
